dynamic time and auto-scroll
