+++ src/index.js
@@ -3,7 +3,6 @@ import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 
 import "bootstrap/dist/css/bootstrap.min.css";
-import "./index.css";
 import App from "./App";
 
 import store from "./redux/store";
